requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
supabase-py
jose